refactor: corregir e reestruturar os testes
