+++ test/aluno/aluno.controller.spec.ts
@@ -1,11 +1,11 @@
+import { AlunoController } from '@/src/aluno/aluno.controller';
+import { AlunoService } from '@/src/aluno/aluno.service';
+import { UnidadeEnum } from '@/src/enum/enumCampus';
+import { CursosEnum } from '@/src/enum/enumCursos';
+import { PronomesEnum } from '@/src/enum/enumPronomes';
+import AuthenticatedRequest from '@/src/types/authenticated-request.interface';
 import { NotFoundException } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
-import { AlunoController } from '../src/aluno/aluno.controller';
-import { AlunoService } from '../src/aluno/aluno.service';
-import { UnidadeEnum } from '../src/enum/enumCampus';
-import { CursosEnum } from '../src/enum/enumCursos';
-import { PronomesEnum } from '../src/enum/enumPronomes';
-import AuthenticatedRequest from '../src/types/authenticated-request.interface';
 
 describe('AlunoController', () => {
   let alunoController: AlunoController;
