+++ test/inscricao/inscricao.service.spec.ts
@@ -1,17 +1,18 @@
+import { EditalEnum } from '@/src/enum/enumEdital';
+import { EnumTipoDocumento } from '@/src/enum/enumTipoDocumento';
+import { StatusDocumento } from '@/src/enum/statusDocumento';
+import { InscricaoService } from '@/src/inscricao/inscricao.service';
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
+import { Aluno } from 'src/entities/aluno/aluno.entity';
+import { Documento } from 'src/entities/documento/documento.entity';
+import { Edital } from 'src/entities/edital/edital.entity';
 import { Pergunta } from 'src/entities/edital/pergunta.entity';
+import { Inscricao } from 'src/entities/inscricao/inscricao.entity';
+import { Validacao } from 'src/entities/validacao/validacao.entity';
 import { EntityManager, Repository } from 'typeorm';
-import { Aluno } from '../src/entities/aluno/aluno.entity';
-import { Documento } from '../src/entities/documento/documento.entity';
-import { Edital } from '../src/entities/edital/edital.entity';
-import { Inscricao } from '../src/entities/inscricao/inscricao.entity';
-import { Validacao } from '../src/entities/validacao/validacao.entity';
-import { EditalEnum } from '../src/enum/enumEdital';
-import { EnumTipoDocumento } from '../src/enum/enumTipoDocumento';
-import { StatusDocumento } from '../src/enum/statusDocumento';
-import { InscricaoService } from '../src/inscricao/inscricao.service';
+
 
 describe('InscricaoService', () => {
   let service: InscricaoService;
